+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## [12.3] - 2022-03-27
+- Hotfix: Italian translation
+- Hotfix: Flush DNS cache now requires confirmation
+- Hotfix: Chinese instead of Taiwanese flag
+- Removed: Pre-made AdBlock lists in Hosts due to several bugs
+
 ## [12.2] - 2022-03-26
 - Hotfix: Crash when loading German language